+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ## [0.129.0] - 2025-03-12
 ### Removed
-- Final version for `svg` export
+- Final version for `svg` export.
 
 ## [0.128.0] - 2025-03-12
 ### Removed
-- Final version for `png` export
+- Final version for `png` export.
 
 ## [0.128.0] - 2025-03-12
 ### Removed